// src/app/models/modal.ts
export interface Modals {
    uid?: string;
    id?: string;
    color?: string;
    userUid?: string;
    userName?: string;
    petName?: string;
    situation?: string;
    age?: string;
    city?: string;
    description?: string;
    likes?: Array<any>;
    dislikes?: Array<any>;
    imageUrl?: string;
    isActive?: boolean;
    date?: string;
}